Asset system progress

+++ Heart/src/Heart/Core/Timing.cpp
@@ -3,6 +3,6 @@
 
 namespace Heart
 {
-    std::unordered_map<std::string, u32> AggregateTimer::s_AggregateTimes = {};
+    std::unordered_map<std::string, double> AggregateTimer::s_AggregateTimes = {};
     std::shared_mutex AggregateTimer::s_Mutex;
 }
